src/ghostbusters_gcp/requirements.txt
# GCP Services
google-cloud-firestore==2.11.1
google-cloud-pubsub==2.18.4
functions-framework==3.4.0

# Basic dependencies
pydantic==2.9.2 